Rely default ConsoleRenderer constructor
This change should allow the default constructor or ConsoleRenderer to
be used and determine if colors should be used of not based on the
presence of colorama as a requirement without having to add it as a
requirment for this library. Should work the same out of the box if
colorama is present but still work without it.